ignite-5732: structure refactoring
