+++ modules/compatibility/src/test/java/org/apache/ignite/compatibility/testframework/junits/CompatibilityTestIgniteNodeRunner.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.testframework.junits.campatibility;
+package org.apache.ignite.compatibility.testframework.junits;
 
 import com.thoughtworks.xstream.XStream;
 import java.io.BufferedReader;
