+++ modules/compatibility/src/test/resources/META-INF/services/org.apache.ignite.plugin.PluginProvider
@@ -1,1 +1,1 @@
-org.apache.ignite.tests.compatibility.plugins.TestCompatibilityPluginProvider
+org.apache.ignite.compatibility.testframework.plugins.TestCompatibilityPluginProvider
